Added links to Pre-Reads and updated the code
