+++ Class-2/README.md
@@ -1,0 +1,22 @@
+#Class-2
+
+## Topics Covered
+
+1. Types of Machine Learning.
+2. Features, training data set, testing data set, input, output etc
+3. Perceptron Learning Algorithm for binary classification
+4. Decision Trees
+5. Introduction to scikit-learn
+
+## Pre-Reads
+Go through the following links
+
+- Machine Learning and its types: [Week 1 of Machine Learning course by Andrew Ng](https://www.coursera.org/learn/machine-learning/)
+- Perceptron and types of Machine Learning: [Lecture 1 of Learning from Data by Yaser Abu Mostafa](https://www.youtube.com/watch?v=VeKeFIepJBU&feature=youtu.be&list=PLCA2C1469EA777F9A&t=1420)
+- Decision Trees: [Video 2 of Machine Learning Recipes by Google](https://www.youtube.com/watch?v=tNa99PG8hR8&list=PLT6elRN3Aer7ncFlaCz8Zz-4B5cnsrOMt)
+
+## Advance Reads
+Will be added shortly
+
+
+
